+++ src/program/GameLoop.cpp
@@ -653,7 +653,6 @@ void GameLoop::sleepSendPreview()
 
 void GameLoop::processInputs(AllInputs &ai)
 {
-    bool modified_by_lua = false;
     ai.clear();
 
     /* Don't record inputs if we are quitting */
@@ -717,8 +716,9 @@ void GameLoop::processInputs(AllInputs &ai)
             }
 
             /* Call lua onInput() here so that a script can modify inputs */
-            Lua::Input::registerInputs(&ai, &modified_by_lua);
+            Lua::Input::registerInputs(&ai);
             Lua::Callbacks::call(Lua::NamedLuaFunction::CallbackInput);
+            Lua::Input::clearInputs();
 
             if (context->config.sc.recording == SharedConfig::RECORDING_WRITE) {
                 /* If the input editor is visible, we should keep future inputs.
@@ -768,8 +768,9 @@ void GameLoop::processInputs(AllInputs &ai)
                 ai = movie.inputs->getInputs();
 
                 /* Allow lua to modify movie inputs even in playback mode */
-                Lua::Input::registerInputs(&ai, &modified_by_lua);
+                Lua::Input::registerInputs(&ai);
                 Lua::Callbacks::call(Lua::NamedLuaFunction::CallbackInput);
+                bool modified_by_lua = Lua::Input::clearInputs();
                 if (modified_by_lua) {
                     movie.inputs->setInputs(ai, true);
                 }
@@ -809,8 +810,9 @@ void GameLoop::processInputs(AllInputs &ai)
                 ai.clear();
 
                 /* Allow lua to modify inputs past the end of the movie */
-                Lua::Input::registerInputs(&ai, &modified_by_lua);
+                Lua::Input::registerInputs(&ai);
                 Lua::Callbacks::call(Lua::NamedLuaFunction::CallbackInput);
+                Lua::Input::clearInputs();
             }
 
             /* Update controller inputs if controller window is shown */